Update Makefile - remove mondo-clingen-review entry (#29)
* Update Makefile - remove mondo-clingen-review entry

The current entry for mondo-clingen-review was set-up to use github branch `issue-9178`. This review work is now complete and the github branch is deleted. Sabrina said "the ols version of branch referring to issue-9178 can be deleted now." in this [Slack comment](https://monarchinitiative.slack.com/archives/G01JACCEDAB/p1755026006723549?thread_ts=1755025726.427119&cid=G01JACCEDAB).

I have commented out the entry so it can more easily be used as a stub in the future if a new mondo-clingen-review entry is added for a future github feature branch.

* Update Makefile

remove mondo-clingen-review entry